src/requirements.txt
accelerate==1.11.0
altair==5.5.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chroma-hnswlib==0.7.1
chromadb==0.4.0
click==8.3.0
coloredlogs==15.0.1
distro==1.9.0
fastapi==0.99.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
groq==0.33.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.10.1
networkx==3.5
nltk==3.9.2
numpy==2.3.4
onnxruntime==1.23.2
openai==2.6.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==12.0.0
posthog==6.7.11
protobuf==6.33.0
psutil==7.1.2
pulsar-client==3.8.0
pyarrow==21.0.0
pydantic==1.10.24
pydeck==0.9.1
PyMuPDF==1.26.5
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
